Rename `EnrichmentHelper` to `EventEnrichments` and move to `.q.projection` package.
