+++ api-java/src/main/java/io/spine/examples/todolist/q/projection/MyListViewProjection.java
@@ -40,7 +40,7 @@ import io.spine.server.projection.Projection;
 import java.util.ArrayList;
 import java.util.List;
 
-import static io.spine.examples.todolist.EnrichmentHelper.getEnrichment;
+import static io.spine.examples.todolist.q.projection.EventEnrichments.getEnrichment;
 import static io.spine.examples.todolist.q.projection.Projections.newTaskListView;
 import static io.spine.examples.todolist.q.projection.Projections.removeViewsByTaskId;
 import static io.spine.examples.todolist.q.projection.Projections.updateTaskItemList;
